requirements.txt
pyyaml>=5.0
watchdog>=2.0
schedule>=1.0
dataclasses;python_version<"3.7"
flask>=2.0
